Update regexps for title parsing.
In particular, directors and writers were not being parsed.

This change also updates the test to be a little more robust in the face
of legitimate changes on the site (e.g. changing ratings and updated
actor ordering).

Most of the poster content URLs seem to have changed since the test file
was last updated.  It may be better to support regexps for string fields
in diffStruct.  For now I just updated the URLs.

All of the tests pass with this change.